client: Move tag feature in a seperate class
