+++ src/index.ts
@@ -1,4 +1,6 @@
 export * from './client';
+export * from './lang-utils';
+export * from './lock-fields-manager';
 export * from './lock';
 export * from './models';
 export * from './mongo-read-stream';
